+++ Sources/MapboxSearch/PublicAPI/MapboxSearchVersion.swift
@@ -1,2 +1,2 @@
 /// Mapbox Search SDK version variable
-public let mapboxSearchSDKVersion = "1.0.0-rc.3"
+public let mapboxSearchSDKVersion = "1.0.0-rc.4"